--- apps/user-web/src/app/(backable)/test/page.tsx
@@ -1,68 +1,0 @@
-'use client';
-
-import { BottomSheet } from '@components/common';
-import useBottomSheet from '@hooks/use-bottom-sheet';
-import { Input } from '@swifty/ui';
-import cn from 'clsx';
-import type { ComponentPropsWithoutRef } from 'react';
-import { useRef, useState } from 'react';
-
-export default function Page() {
-  const [isOpen, open, close] = useBottomSheet();
-  const [value, setValue] = useState('');
-  const sex = ['남성', '여성'] as const;
-  const ref = useRef<HTMLInputElement>(null);
-
-  const onOpen = () => {
-    open();
-    ref.current?.focus();
-  };
-
-  const onSelectValue = (value: string) => {
-    setValue(value);
-    close();
-  };
-  return (
-    <div>
-      <div onClick={onOpen}>
-        <Input
-          name="성별"
-          placeholder="성별"
-          defaultValue={value}
-          ref={ref}
-          type="button"
-        />
-      </div>
-
-      <BottomSheet isOpen={isOpen} onDismiss={close}>
-        {sex.map((item, index) => (
-          <Item
-            key={item}
-            value={item}
-            underline={index !== sex.length - 1}
-            onClick={() => onSelectValue(item)}
-          />
-        ))}
-      </BottomSheet>
-    </div>
-  );
-}
-
-type Item = {
-  value: string;
-  underline?: boolean;
-} & ComponentPropsWithoutRef<'div'>;
-
-function Item({ value, underline, ...props }: Item) {
-  return (
-    <div
-      className={cn(
-        'bg-black text-white text-center text-16 font-medium py-[22px] hover:bg-gray-800 focus:bg-swifty-gray-700 focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-swifty-gray-500',
-        underline && 'border-swifty-color-700 border-b-2 border-solid',
-      )}
-      {...props}
-    >
-      {value}
-    </div>
-  );
-}
