+++ apps/user-web/src/components/common/index.ts
@@ -14,5 +14,5 @@ export { default as MenuTiles } from './menu-tiles';
 export { default as ImageWithFallback } from './image-with-fallback';
 export { default as FadeOverlay } from './fade-overlay';
 export { default as Overlay } from './overlay';
-export { default as Select } from './select';
+export { default as BottomSelect } from './select';
 export * from './container';
